Plant-Disease-identification/requirements.txt
streamlit
tensorflow==2.13.0
keras
numpy
pandas
Pillow
matplotlib
